Refactor availability screen components
